Lets try this again.
