+++ src/utilities/debugLogger.js
@@ -1,0 +1,15 @@
+/**
+ * Logs out a message or calls debug if env variable is set true.
+ * @param message
+ * @param debug
+ */
+export const logOutMessageOrDebug = ({ message, debug }) => {
+  if (process.env.SHOW_LOGS === 'true') {
+    if (debug !== false) {
+      debug();
+    }
+    if (message !== undefined) {
+      console.log(message);
+    }
+  }
+};
