+++ src/utilities/test-utils.js
@@ -114,21 +114,6 @@ export const getNodesByType = (c, type) => {
   return toReturn;
 };
 
-/**
- * Logs out a message or calls debug if env variable is set true.
- * @param message
- * @param debug
- */
-export const logOutMessageOrDebug = ({ message, debug }) => {
-  if (process.env.SHOW_LOGS === 'true') {
-    if (debug !== false) {
-      debug();
-    }
-    if (message !== undefined) {
-      console.log(message);
-    }
-  }
-};
 
 export * from '@testing-library/react';
